Created mealRouter and scan route.

+++ apps/server/routers/mealRouter.ts
@@ -4,7 +4,15 @@ import { InternalServerError } from '../classes/Error'
 import multer from 'multer'
 import fs from 'fs/promises' // For reading files asynchronously
 import path from 'path'
+import dotenv from 'dotenv'
 
+import downloadFile from '../utils/downloadFile'
+import { CaloriesClaculatorGPT } from '../services/caloriesCalculatorGpt.service'
+import { CaloriesClaculator } from '../services/caloriesCalculator.service'
+
+dotenv.config()
+
+const calculationMethod = process.env.CURRENT_CALCULATOR_SECRET
 const upload = multer({ dest: 'uploads/' })
 const mealRouter = express.Router()
 
@@ -13,6 +21,22 @@ function generateFileName(originalName: string): string {
   const extension = path.extname(originalName)
   const baseName = path.basename(originalName, extension)
   return `${baseName}-${timestamp}${extension}`
+}
+
+function appendExtensionBasedOnMimeType(
+  fileName: string,
+  mimeType: string
+): string {
+  const mimeTypes: { [key: string]: string } = {
+    'image/jpeg': '.jpg',
+    'image/png': '.png',
+    'image/gif': '.gif',
+    'application/pdf': '.pdf'
+    // Add more mime types and their corresponding extensions as needed
+  }
+
+  const extension = mimeTypes[mimeType] || ''
+  return fileName + extension
 }
 
 mealRouter.post(
@@ -30,10 +54,18 @@ mealRouter.post(
       const fileBuffer = await fs.readFile(file.path)
       const supaBaseClient = initSupaBaseClient()
 
+      console.log('File:', file)
+
+      const fileName = generateFileName(file.filename)
+
+      const fileNameWithExtension = appendExtensionBasedOnMimeType(
+        fileName,
+        file.mimetype
+      )
       // Upload the file to Supabase Storage
-      const { data, error } = await supaBaseClient.storage
-        .from('snackntrack')
-        .upload(`public/${generateFileName(file.filename)}`, fileBuffer, {
+      const { error } = await supaBaseClient.storage
+        .from('snack-n-track-bucket')
+        .upload(`meal/${fileNameWithExtension}`, fileBuffer, {
           contentType: file.mimetype
         })
 
@@ -44,10 +76,28 @@ mealRouter.post(
         )
       }
 
+      const { data: publicURL } = await supaBaseClient.storage
+        .from('snack-n-track-bucket')
+        .getPublicUrl(`meal/${fileNameWithExtension}`)
+
+      if (!publicURL) {
+        throw new InternalServerError('Failed to get public URL')
+      }
+
       // Delete the file from the local storage after successful upload
       await fs.unlink(file.path)
+      console.log(`fileUrl: ${publicURL.publicUrl}`)
 
-      res.status(200).json({ data })
+      const imagePath = path.resolve(__dirname, '../downloads')
+      downloadFile(publicURL.publicUrl, `../downloads/${fileNameWithExtension}`)
+
+      const caloriesCalculator = new CaloriesClaculator(CaloriesClaculatorGPT)
+      const calories = await caloriesCalculator.calculateCalories(
+        `../downloads/${fileNameWithExtension}`
+      )
+      console.log('Calories:', calories)
+
+      res.status(200).json({ calories })
     } catch (error: any) {
       console.error('Upload error:', error)
       res
